Move JournalLock class into its own file
Closes #891
