+++ EDMarketConnector.py
@@ -39,7 +39,7 @@ from config import appversion, appversion_nobuild, config, copyright
 # isort: on
 
 from EDMCLogging import edmclogger, logger, logging
-from monitor import JournalLock
+from journal_lock import JournalLock
 
 if __name__ == '__main__':  # noqa: C901
     # Command-line arguments
